remove plurals data sources
