--- nutanix/data_source_nutanix_images_test.go
@@ -1,52 +1,0 @@
-package nutanix
-
-import (
-	"fmt"
-	"testing"
-
-	"github.com/hashicorp/terraform/helper/acctest"
-	"github.com/hashicorp/terraform/helper/resource"
-	"github.com/hashicorp/terraform/terraform"
-)
-
-func TestAccNutanixImagesDataSource_basic(t *testing.T) {
-
-	resource.Test(t, resource.TestCase{
-		PreCheck:  func() { testAccPreCheck(t) },
-		Providers: testAccProviders,
-		Steps: []resource.TestStep{
-			{
-				Config: testAccImagesDataSourceConfig(acctest.RandIntRange(0, 500)),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNutanixImagesExists("data.nutanix_images.test1"),
-				),
-			},
-		},
-	})
-}
-
-func testAccCheckNutanixImagesExists(n string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[n]
-		if !ok {
-			return fmt.Errorf("not found: %s", n)
-		}
-		if rs.Primary.ID == "" {
-			return fmt.Errorf("no ID is set")
-		}
-
-		return nil
-	}
-}
-
-func testAccImagesDataSourceConfig(rNumber int) string {
-	return fmt.Sprintf(`
-resource "nutanix_image" "test" {
-  name        = "Ubuntu-%d"
-  description = "Ubuntu"
-  source_uri  = "http://archive.ubuntu.com/ubuntu/dists/bionic/main/installer-amd64/current/images/netboot/mini.iso"
-}
-
-
-data "nutanix_images" "test1" {}`, rNumber)
-}
