+++ nutanix/provider.go
@@ -47,16 +47,12 @@ func Provider() terraform.ResourceProvider {
 			},
 		},
 		DataSourcesMap: map[string]*schema.Resource{
-			"nutanix_virtual_machine":        dataSourceNutanixVirtualMachine(),
-			"nutanix_virtual_machines":       dataSourceNutanixVirtualMachines(),
-			"nutanix_image":                  dataSourceNutanixImage(),
-			"nutanix_images":                 dataSourceNutanixImages(),
-			"nutanix_subnet":                 dataSourceNutanixSubnet(),
-			"nutanix_subnets":                dataSourceNutanixSubnets(),
-			"nutanix_clusters":               dataSourceNutanixClusters(),
-			"nutanix_network_security_rule":  dataSourceNutanixNetworkSecurityRule(),
-			"nutanix_network_security_rules": dataSourceNutanixNetworkSecurityRules(),
-			"nutanix_cluster":                dataSourceNutanixCluster(),
+			"nutanix_image":                 dataSourceNutanixImage(),
+			"nutanix_subnet":                dataSourceNutanixSubnet(),
+			"nutanix_cluster":               dataSourceNutanixCluster(),
+			"nutanix_clusters":              dataSourceNutanixClusters(),
+			"nutanix_virtual_machine":       dataSourceNutanixVirtualMachine(),
+			"nutanix_network_security_rule": dataSourceNutanixNetworkSecurityRule(),
 			// "nutanix_volume_group":           dataSourceNutanixVolumeGroup(),
 			// "nutanix_volume_groups":          dataSourceNutanixVolumeGroups(),
 		},
