--- website/docs/d/images.html.markdown
@@ -1,102 +1,0 @@
----
-layout: "nutanix"
-page_title: "NUTANIX: nutanix_images"
-sidebar_current: "docs-nutanix-datasource-images"
-description: |-
- Describes a List of Images
----
-
-# nutanix_images
-
-Describes a List of Images
-
-## Example Usage
-
-```hcl
-resource "nutanix_image" "ubuntu" {
-  name        = "Ubuntu"
-  description = "Ubuntu"
-  source_uri  = "http://archive.ubuntu.com/ubuntu/dists/bionic/main/installer-amd64/current/images/netboot/mini.iso"
-}
-
-resource "nutanix_image" "centos" {
-  name        = "Centos"
-  description = "Centos"
-  source_uri  = "http://earth.corp.nutanix.com/isos/linux/centos/7/CentOS-7.3-x86_64-Minimal-1611.iso""
-}
-
-#list all images
-data "nutanix_images" "test" {}
-```
-
-## Argument Reference
-
-The following arguments are supported:
-
-* `metadata`: Represents virtual machine UUID
-
-### Metadata Argument
-
-The metadata attribute supports the following:
-
-* `kind`: - The kind name.
-* `sort_attribute`: The attribute to perform sort on.
-* `filter`: - The filter in FIQL syntax used for the results.
-* `length`: - The number of records to retrieve relative to the offset.
-* `sort_order`: - The sort order in which results are returned
-* `offset`: - Offset from the start of the entity list
-
-## Attribute Reference
-
-The following attributes are exported:
-
-* `entities`: - A list of virtual machines.
-
-### Entities Attribute
-
-The entities attribute supports the following:
-
-* `name`: -  The name for the image.
-* `categories`: - Categories for the image.
-* `project_reference`: - The reference to a project.
-* `owner_reference`: - The reference to a user.
-* `availability_zone_reference`: - The reference to a availability_zone.
-* `api_version` -
-* `description`: - A description for image.
-* `architecture`: - The supported CPU architecture for a disk image.
-* `checksum`: - Checksum of the image. The checksum is used for image validation if the image has a source specified. For images that do not have their source specified the checksum is generated by the image service.
-* `image_type`: - The type of image.
-* `source_uri`: - The source URI points at the location of the source image which is used to create/update image.
-* `version`: - The image version.
-* `metadata`: - The image kind metadata.
-* `retrieval_uri_list`: - List of URIs where the raw image data can be accessed.
-* `size_bytes`: - The size of the image in bytes.
-* `state`: - The state of the image.
-
-### Metadata
-
-The metadata attribute exports the following:
-
-* `last_update_time`: - UTC date and time in RFC-3339 format when image was last updated.
-* `UUID`: - image UUID.
-* `creation_time`: - UTC date and time in RFC-3339 format when image was created.
-* `spec_version`: - Version number of the latest spec.
-* `spec_hash`: - Hash of the spec. This will be returned from server.
-* `name`: - image name.
-
-### Reference
-
-The `project_reference`, `owner_reference`, `availability_zone_reference`, `cluster_reference`, attributes supports the following:
-
-* `kind`: - The kind name (Default value: project).
-* `name`: - the name.
-* `UUID`: - the UUID.
-
-### Version
-
-The version attribute supports the following:
-
-* `product_name`: - Name of the producer/distribution of the image. For example windows or red hat.
-* `product_version`: - Version string for the disk image.
-
-See detailed information in [Nutanix Image](http://developer.nutanix.com/reference/prism_central/v3/#images).
